Add support to replace scitoken with macaroon on redirects
This commit adds a new configuration directive: 'http.redirtoken'.

When this new directive is enabled, a valid SciToken in the URL as
an `authz` argument will be replaced on redirection with a sufficient
Macaroon instead. The redirected Macaroon will have a short lifetime
and be specific to the path being accessed.